- Move components to src folder - Solve the fetch to local - Add env.js configuration
