+++ src/config.js
@@ -1,0 +1,4 @@
+const localAddress = '172.13.3.70' // Local ip
+
+
+export {localAddress}
